% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmode.R
\name{getmode}
\alias{getmode}
\title{Mode function}
\usage{
getmode(x)
}
\arguments{
\item{x}{Vecor, data frame, or array}
}
\value{
the mode of that data
}
\description{
More detailed description
}
\examples{
\dontrun{getmode(c(12, 18, 18, 24))}

}
