% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATH4753ayco0000.R
\docType{package}
\name{MATH4753ayco0000}
\alias{MATH4753ayco0000}
\title{Lab Functions Package}
\description{
Contains functions used in MATH4753 labs
}
\author{
David Aycock \email{david.b.aycock-1@ou.edu}
}
