% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{bootstrap}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of times sampled}

\item{x}{the sample vector}

\item{fun}{function being applied}

\item{alpha}{confidence interval (alpha,1-alpha)}

\item{cx}{character size on histogram}

\item{...}{any other parameter to customize the histogram}
}
\value{
A histogram that had a line that indicates the statistic with text of the value of the statistic alongside text at the location where the bootstrap interval is
}
\description{
bootstrap
}
\examples{
\dontrun{x=myboot(ddt$DDT)}
}
