% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{my confidence interval}
\usage{
myci(x, conf.int = 0.95)
}
\arguments{
\item{x}{vector of data}

\item{conf.int}{the confidence interval, defaulted to 95%}
}
\value{
a confidence interval for the mean of the data given in x
}
\description{
my confidence interval
}
\examples{
\dontrun{set.seed(30); x=rnorm(30,mean=10,sd=12); myci(x)}
}
