% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Poison distribution using CLT}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{number of random values}

\item{iter}{number if iterations}

\item{lambda}{the mean}
}
\value{
A histogram, barplot and probability chart
}
\description{
Poison distribution using CLT
}
\examples{
\dontrun{mycltp(n=20,iter=10000)}
}
