% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{maximum likelihood}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{log function of sample}

\item{x}{sample vector}

\item{param}{the parameter being calculated}

\item{...}{any parameters that can customize the plot}
}
\value{
a graph that draws a line through the value of the maximum likelihood
}
\description{
maximum likelihood
}
\examples{
\dontrun{mymaxlik(lfun = function(x,param) log(dbinom(x,prob=param,size=20)), x = c(3,3,4,3,4,5,5,4), param = seq(0,1,length=1000))}
}
