% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{normal curve maker}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{x-value that will be shaded up to from -inf}
}
\value{
curve with shaded region from -inf to a and the area under the curve from -inf to a
}
\description{
normal curve maker
}
\examples{
\dontrun{myncurve(10,4,1)}
}
