% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{Random Sample barplot generator}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{size of the sample}

\item{iter}{the iterator}

\item{time}{time between each execution}
}
\value{
iter amount of barplots for iter samples
}
\description{
Random Sample barplot generator
}
\examples{
mysample(10, 10, 1)
}
